refactor(crypto): consolidate mint-quote signing in NUT20, amended pair internal
cashubtc/nuts#375 defines one mint-quote message for NUT-20 single and NUT-29
batch minting, so the NUT29 module dissolves into NUT20.ts. On v4 the released
signMintQuote/verifyMintQuoteSignature keep their legacy bytes and the amended
pair (signMintQuoteAmended/verifyMintQuoteSignatureAmended) stays out of the
public barrel; v5 exports the amended pair as signMintQuote directly. Schnorr
plumbing now reuses core.ts (new schnorrVerifyDigest counterpart).
